pyproject.toml
[tool.black]
line-length = 120

[tool.poetry]
name = "fastapi-scratch"
version = "0.1.0"
description = ""
authors = ["Pryann <pryann@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
bcrypt = "^4.1.3"
humps = "^0.2.2"
pydantic-settings = "^2.3.1"
asyncpg = "^0.29.0"
alembic = "^1.13.1"

pyjwt = "^2.8.0"
slowapi = "^0.1.9"
aioredis = "^2.0.1"
loguru = "^0.7.2"
sentry-sdk = "^2.6.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "src.scripts:dev"
prod = "src.scripts:prod"
alembic-init = "src.scripts:alembic_init"
alembic-revision = "src.scripts:alembic_revision"
alembic-upgrade = "src.scripts:alembic_upgrade"
alembic-downgrade = "src.scripts:alembic_downgrade"

[[tool.poetry.packages]]
include = "**/*.py"